7.1.5 Penalty Charge Updates

+++ idas.chox.web/src/main/java/idas/chox/web/actions/InvoiceDetailAction.java
@@ -1262,7 +1262,7 @@ public class InvoiceDetailAction extends BaseAction implements Preparable {
         }
     }
 
-    public long getInvoicedDays() {
+    public int getInvoicedDays() {
         // long dateDiff = DateHelper.getNumberOf24HourPeriodsBetween(getDateInvoiced(), new Date()) + 1;
         return invoice.getInvoicedDays();
